uniform copy button / optional title on CodeGroup
